Sort messages added to Github for check consistency
To know if a message should be added to Github we check if it was
already added.
Or if we have more than one messages for a line, they are not guaranteed
to be in the same order, so the check fail.

This commit simply sort the lines before adding them to the final
message to check.

(It also simplifies this part of the code using the fact that `set.add`
will not add a new occurence of the value if already present, so the
extra set and the `if..in` are not needed)